Refactor common pattern to implement iterators
We can share code between some of the word, grapheme clusters, and sentence
iterators since their differences are only the payload.
