+++ experimental/segmenter/src/grapheme.rs
@@ -3,7 +3,6 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 use alloc::vec::Vec;
-use core::str::CharIndices;
 use icu_provider::prelude::*;
 
 use crate::complex::{Dictionary, LstmPayloads};
@@ -13,20 +12,18 @@ use crate::{provider::*, SegmenterError};
 use utf8_iter::Utf8CharIndices;
 
 /// Grapheme cluster break iterator for an `str` (a UTF-8 string).
-pub type GraphemeClusterBreakIteratorUtf8<'l, 's> =
-    RuleBreakIterator<'l, 's, GraphemeClusterBreakTypeUtf8>;
+pub type GraphemeClusterBreakIteratorUtf8<'l, 's> = RuleBreakIterator<'l, 's, RuleBreakTypeUtf8>;
 
 /// Grapheme cluster break iterator for a potentially invalid UTF-8 string.
 pub type GraphemeClusterBreakIteratorPotentiallyIllFormedUtf8<'l, 's> =
-    RuleBreakIterator<'l, 's, GraphemeClusterBreakTypePotentiallyIllFormedUtf8>;
+    RuleBreakIterator<'l, 's, RuleBreakTypePotentiallyIllFormedUtf8>;
 
 /// Grapheme cluster break iterator for a Latin-1 (8-bit) string.
 pub type GraphemeClusterBreakIteratorLatin1<'l, 's> =
-    RuleBreakIterator<'l, 's, GraphemeClusterBreakTypeLatin1>;
+    RuleBreakIterator<'l, 's, RuleBreakTypeLatin1>;
 
 /// Grapheme cluster break iterator for a UTF-16 string.
-pub type GraphemeClusterBreakIteratorUtf16<'l, 's> =
-    RuleBreakIterator<'l, 's, GraphemeClusterBreakTypeUtf16>;
+pub type GraphemeClusterBreakIteratorUtf16<'l, 's> = RuleBreakIterator<'l, 's, RuleBreakTypeUtf16>;
 
 /// Segments a string into grapheme clusters.
 ///
@@ -160,82 +157,5 @@ impl GraphemeClusterSegmenter {
             lstm: &self.lstm,
             grapheme: None,
         }
-    }
-}
-
-pub struct GraphemeClusterBreakTypeUtf8;
-
-impl<'l, 's> RuleBreakType<'l, 's> for GraphemeClusterBreakTypeUtf8 {
-    type IterAttr = CharIndices<'s>;
-    type CharType = char;
-
-    fn get_current_position_character_len(iter: &RuleBreakIterator<Self>) -> usize {
-        iter.get_current_codepoint().len_utf8()
-    }
-
-    fn handle_complex_language(
-        _: &mut RuleBreakIterator<Self>,
-        _: Self::CharType,
-    ) -> Option<usize> {
-        unreachable!()
-    }
-}
-
-pub struct GraphemeClusterBreakTypePotentiallyIllFormedUtf8;
-
-impl<'l, 's> RuleBreakType<'l, 's> for GraphemeClusterBreakTypePotentiallyIllFormedUtf8 {
-    type IterAttr = Utf8CharIndices<'s>;
-    type CharType = char;
-
-    fn get_current_position_character_len(iter: &RuleBreakIterator<Self>) -> usize {
-        iter.get_current_codepoint().len_utf8()
-    }
-
-    fn handle_complex_language(
-        _: &mut RuleBreakIterator<Self>,
-        _: Self::CharType,
-    ) -> Option<usize> {
-        unreachable!()
-    }
-}
-
-pub struct GraphemeClusterBreakTypeLatin1;
-
-impl<'l, 's> RuleBreakType<'l, 's> for GraphemeClusterBreakTypeLatin1 {
-    type IterAttr = Latin1Indices<'s>;
-    type CharType = u8;
-
-    fn get_current_position_character_len(_: &RuleBreakIterator<Self>) -> usize {
-        unreachable!()
-    }
-
-    fn handle_complex_language(
-        _: &mut RuleBreakIterator<Self>,
-        _: Self::CharType,
-    ) -> Option<usize> {
-        unreachable!()
-    }
-}
-
-pub struct GraphemeClusterBreakTypeUtf16;
-
-impl<'l, 's> RuleBreakType<'l, 's> for GraphemeClusterBreakTypeUtf16 {
-    type IterAttr = Utf16Indices<'s>;
-    type CharType = u32;
-
-    fn get_current_position_character_len(iter: &RuleBreakIterator<Self>) -> usize {
-        let ch = iter.get_current_codepoint();
-        if ch >= 0x10000 {
-            2
-        } else {
-            1
-        }
-    }
-
-    fn handle_complex_language(
-        _: &mut RuleBreakIterator<Self>,
-        _: Self::CharType,
-    ) -> Option<usize> {
-        unreachable!()
     }
 }
