+++ experimental/segmenter/src/sentence.rs
@@ -3,7 +3,6 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 use alloc::vec::Vec;
-use core::str::CharIndices;
 use icu_provider::prelude::*;
 
 use crate::complex::{Dictionary, LstmPayloads};
@@ -13,17 +12,17 @@ use crate::{provider::*, SegmenterError};
 use utf8_iter::Utf8CharIndices;
 
 /// Sentence break iterator for an `str` (a UTF-8 string).
-pub type SentenceBreakIteratorUtf8<'l, 's> = RuleBreakIterator<'l, 's, SentenceBreakTypeUtf8>;
+pub type SentenceBreakIteratorUtf8<'l, 's> = RuleBreakIterator<'l, 's, RuleBreakTypeUtf8>;
 
 /// Sentence break iterator for potentially invalid UTF-8 strings
 pub type SentenceBreakIteratorPotentiallyIllFormedUtf8<'l, 's> =
-    RuleBreakIterator<'l, 's, SentenceBreakTypePotentiallyIllFormedUtf8>;
+    RuleBreakIterator<'l, 's, RuleBreakTypePotentiallyIllFormedUtf8>;
 
 /// Sentence break iterator for a Latin-1 (8-bit) string.
-pub type SentenceBreakIteratorLatin1<'l, 's> = RuleBreakIterator<'l, 's, SentenceBreakTypeLatin1>;
+pub type SentenceBreakIteratorLatin1<'l, 's> = RuleBreakIterator<'l, 's, RuleBreakTypeLatin1>;
 
 /// Sentence break iterator for a UTF-16 string.
-pub type SentenceBreakIteratorUtf16<'l, 's> = RuleBreakIterator<'l, 's, SentenceBreakTypeUtf16>;
+pub type SentenceBreakIteratorUtf16<'l, 's> = RuleBreakIterator<'l, 's, RuleBreakTypeUtf16>;
 
 /// Supports loading sentence break data, and creating sentence break iterators for different string
 /// encodings.
@@ -146,81 +145,5 @@ impl SentenceSegmenter {
             lstm: &self.lstm,
             grapheme: None,
         }
-    }
-}
-
-pub struct SentenceBreakTypeUtf8;
-
-impl<'l, 's> RuleBreakType<'l, 's> for SentenceBreakTypeUtf8 {
-    type IterAttr = CharIndices<'s>;
-    type CharType = char;
-
-    fn get_current_position_character_len(iter: &RuleBreakIterator<Self>) -> usize {
-        iter.get_current_codepoint().len_utf8()
-    }
-
-    fn handle_complex_language(
-        _: &mut RuleBreakIterator<Self>,
-        _: Self::CharType,
-    ) -> Option<usize> {
-        unreachable!()
-    }
-}
-pub struct SentenceBreakTypePotentiallyIllFormedUtf8;
-
-impl<'l, 's> RuleBreakType<'l, 's> for SentenceBreakTypePotentiallyIllFormedUtf8 {
-    type IterAttr = Utf8CharIndices<'s>;
-    type CharType = char;
-
-    fn get_current_position_character_len(iter: &RuleBreakIterator<Self>) -> usize {
-        iter.get_current_codepoint().len_utf8()
-    }
-
-    fn handle_complex_language(
-        _: &mut RuleBreakIterator<Self>,
-        _: Self::CharType,
-    ) -> Option<usize> {
-        unreachable!()
-    }
-}
-
-pub struct SentenceBreakTypeLatin1;
-
-impl<'l, 's> RuleBreakType<'l, 's> for SentenceBreakTypeLatin1 {
-    type IterAttr = Latin1Indices<'s>;
-    type CharType = u8;
-
-    fn get_current_position_character_len(_: &RuleBreakIterator<Self>) -> usize {
-        unreachable!()
-    }
-
-    fn handle_complex_language(
-        _: &mut RuleBreakIterator<Self>,
-        _: Self::CharType,
-    ) -> Option<usize> {
-        unreachable!()
-    }
-}
-
-pub struct SentenceBreakTypeUtf16;
-
-impl<'l, 's> RuleBreakType<'l, 's> for SentenceBreakTypeUtf16 {
-    type IterAttr = Utf16Indices<'s>;
-    type CharType = u32;
-
-    fn get_current_position_character_len(iter: &RuleBreakIterator<Self>) -> usize {
-        let ch = iter.get_current_codepoint();
-        if ch >= 0x10000 {
-            2
-        } else {
-            1
-        }
-    }
-
-    fn handle_complex_language(
-        _: &mut RuleBreakIterator<Self>,
-        _: Self::CharType,
-    ) -> Option<usize> {
-        unreachable!()
     }
 }
